% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{aggregate_category}
\alias{aggregate_category}
\title{Aggregate count by category}
\usage{
aggregate_category(
  locations,
  focal = "realm",
  group_by = "taxon",
  return_frequency = FALSE
)
}
\arguments{
\item{locations}{Data.frame of locations (i.e., from locality_add_habitat_biome)}

\item{focal}{Column name to aggregate data over.}

\item{group_by}{What column name to use for grouping}

\item{return_frequency}{Boolean; if TRUE, give frequency, not counts}
}
\value{
A matrix of counts or frequencies
}
\description{
Get count (or frequency) of number of entries per taxon for each category. The return will be a matrix with rows equal to your taxa and columns all possible categories for the focal column, with entries being the number / frequency of records for that taxon for that category
}
\examples{
locations <- spocc_taxon_query("Bubo", limit=500)
locations <- locality_clean(locations)
locations <- locality_add_habitat_biome(locations)
biome_counts <- aggregate_category(locations, focal="biome", group_by="taxon")
print(head(biome_counts))
realm_frequencies <- aggregate_category(locations, focal="realm", group_by="taxon", return_frequency=TRUE)
print(head(realm_frequencies))
}
