% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{chapter2_drop_type}
\alias{chapter2_drop_type}
\title{Return a chapter2 object with only the selected kind of traits}
\usage{
chapter2_drop_type(chapter2, keep = c("continuous", "discrete"))
}
\arguments{
\item{chapter2}{a chapter2 class object, with tree and data}

\item{keep}{Whether you want to keep continuous or discrete data}
}
\value{
A chapter2 object with the tree and only the chosen kind of data
}
\description{
Return a chapter2 object with only the selected kind of traits
}
