% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{chapter2_fitContinuous}
\alias{chapter2_fitContinuous}
\title{Try geiger's fitContinuous on all the continuous datasets using all geiger models}
\usage{
chapter2_fitContinuous(chapter2, models = c("BM", "OU", "EB", "trend",
  "lambda", "kappa", "delta", "drift", "white"), ncores = NULL)
}
\arguments{
\item{chapter2}{a chapter2 class object, with tree and data}

\item{models}{which models to use}

\item{ncores}{how many cores to use; if NULL, detects automatically}
}
\description{
Note that this ignores SE at the moment. If a character is missing information for a taxon, that taxon is deleted for just that character only.
}
\examples{
data(geospiza,package="geiger")
chapter2 <- match_data(geospiza$phy, geospiza$dat)
results <- chapter2_fitContinuous(chapter2)
# Look at model for OU for character 1
print(results[["OU"]][["wingL"]])
}
