% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{chapter2_fitGeiger}
\alias{chapter2_fitGeiger}
\title{Try geiger's fitContinuous or fitDiscrete on all the continuous datasets using all geiger models}
\usage{
chapter2_fitGeiger(
  chapter2,
  models = NULL,
  keep = c("continuous", "discrete"),
  ncores = NULL
)
}
\arguments{
\item{chapter2}{a chapter2 class object, with tree and data}

\item{models}{which models to use. If NULL, uses all available}

\item{keep}{continuous or discrete data}

\item{ncores}{how many cores to use; if NULL, detects automatically}
}
\value{
A two dimensional list. The first dimension is model, the second is character
}
\description{
Note that this ignores SE at the moment. If a character is missing information for a taxon, that taxon is deleted for just that character only.
}
\examples{
data(geospiza,package="geiger")
chapter2 <- match_data(geospiza$phy, geospiza$dat)
results <- chapter2_fitContinuous(chapter2)
# Look at model for OU for character 1
print(results[["OU"]][["wingL"]])
}
