% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{check_numeric_vector}
\alias{check_numeric_vector}
\title{Check a vector to see if the elements are numeric}
\usage{
check_numeric_vector(x)
}
\arguments{
\item{x}{Vector of class whatever}
}
\value{
A boolean on whether the elements, other than NAs, are numeric
}
\description{
Check a vector to see if the elements are numeric
}
