% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{contrasts_correlations}
\alias{contrasts_correlations}
\title{All pairwise correlations}
\usage{
contrasts_correlations(phy, traits)
}
\arguments{
\item{phy}{A phylo object}

\item{traits}{A data.frame of traits, with taxon names as rownames}
}
\value{
A list with three objects: the correlations, lower, and upper.
}
\description{
Use independent contrasts on each pair of traits. If a trait is missing for a taxon, drop that taxon for all analyses with that trait but not others. It runs cor.test on the positivized contrasts and returns the correlation, lower, and upper 95% values for the contrasts
}
