% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{convert_continuous_to_discrete}
\alias{convert_continuous_to_discrete}
\title{Continuous to discrete}
\usage{
convert_continuous_to_discrete(latlon)
}
\arguments{
\item{latlon}{The data.frame of points, with species, latitude, and longitude columns}
}
\value{
A data.frame as above but with a region column added
}
\description{
Convert a set of latitudes and longitudes to biogeographic regions
}
