% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{descendant_species}
\alias{descendant_species}
\title{Get all descendant species of the taxon}
\usage{
descendant_species(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
vector of species names
}
\description{
Uses taxize and GBIF. Note it can only get up to 1000 names currently
}
