% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoldata.R
\name{eol_data}
\alias{eol_data}
\title{Query Encyclopedia of Life with a URL to return all data for a given taxon}
\usage{
eol_data(species)
}
\arguments{
\item{species}{Species name to query}
}
\value{
A data frame of trait category, value, source, and definitions
}
\description{
Query Encyclopedia of Life with a URL to return all data for a given taxon
}
