% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{fishbase_genus_query}
\alias{fishbase_genus_query}
\title{Get information on fish localities rfishbase}
\usage{
fishbase_genus_query(genus)
}
\arguments{
\item{genus}{fish genus name}
}
\value{
a data frame of SpecCode, Species, NorthernLatitude, NorthernLatitudeNS, SouthernLatitude, SouthernLatitudeNS, WesternLongitude, WesternLongitudeEW, EasternLongitude, EasternLongitudeEW
}
\description{
Get information on fish localities rfishbase
}
