% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{fishbase_species_query}
\alias{fishbase_species_query}
\title{Get information on fish localities rfishbase}
\usage{
fishbase_species_query(species)
}
\arguments{
\item{species}{vector of fish species names}
}
\value{
a data frame of SpecCode, Species, NorthernLatitude, NorthernLatitudeNS, SouthernLatitude, SouthernLatitudeNS, WesternLongitude, WesternLongitudeEW, EasternLongitude, EasternLongitudeEW
}
\description{
Get information on fish localities rfishbase
}
