% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_datelife_biggest}
\alias{get_datelife_biggest}
\title{Get biggest tree from datelife}
\usage{
get_datelife_biggest(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
phylo object
}
\description{
Get biggest tree from datelife
}
