% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_eol}
\alias{get_eol}
\title{Get info from Encyclopedia of Life}
\usage{
get_eol(taxon)
}
\arguments{
\item{taxon}{The clade to investigate}
}
\value{
data.frame with species name, trait category, trait value, and other information
}
\description{
Get info from Encyclopedia of Life
}
\examples{
info <- get_eol("Gallus")
}
