% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_genbank_count}
\alias{get_genbank_count}
\title{Information about the number of species in the clade in genbank}
\usage{
get_genbank_count(taxon)
}
\arguments{
\item{taxon}{The clade of interest}
}
\value{
The count of species in genbank
}
\description{
Information about the number of species in the clade in genbank
}
\examples{
taxon <- "Myrmecocystus"
print(paste("There are", get_genbank_count(taxon), "species of", taxon, "in GenBank"))
}
