% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_genbank_count_by_gene}
\alias{get_genbank_count_by_gene}
\title{Get the count of number of sequences for different genes for this taxon}
\usage{
get_genbank_count_by_gene(
  taxon,
  focal.genes = c("COI", "18S", "28S", "matk", "rbcl")
)
}
\arguments{
\item{taxon}{The clade of interest}

\item{focal.genes}{Vector of gene names}
}
\value{
vector of counts by gene, with label being the given gene name
}
\description{
This will use a few popular genes by default, but you can pass your own instead.
}
\examples{
get_genbank_count_by_gene("Myrmecocystus")
}
