% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_otol}
\alias{get_otol}
\title{Get info from Open Tree of Life}
\usage{
get_otol(taxon)
}
\arguments{
\item{taxon}{The clade to investigate}
}
\value{
list containing studies (a data.frame of info about that taxon. Columns include year of the study, number of taxa in the tree from that study, the study citation, and the study DOI) and ntaxa (the number of taxa in OpenTree for that taxon).
}
\description{
Get info from Open Tree of Life
}
\examples{
info <- get_otol("Gallus")
histogram(info$studies$year) # Years in which chicken papers in OpenTree were published
}
