% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_pubmed}
\alias{get_pubmed}
\title{Get information on the clade from pubmed}
\usage{
get_pubmed(taxon, search.string = " AND phylogeny", retmax = 50)
}
\arguments{
\item{taxon}{The clade of interest}

\item{search.string}{Include spaces and AND and similar search elements}

\item{retmax}{how many papers to return}
}
\value{
List with a count element (integer) and a recent.papers element (data.frame)
}
\description{
This will give the total number of pubmed articles mentioning the taxon name and other search string and information on the most recent such papers.
}
\examples{
taxon <- "Formicidae"
results <- get_pubmed(taxon)
print(paste("There are", results$count, "papers on", taxon, "and phylogeny"))
print(results$recent.papers[,c("Date", "FirstAuthor")])
}
