% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_wikipedia_summary}
\alias{get_wikipedia_summary}
\title{Get Wikipedia summary}
\usage{
get_wikipedia_summary(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
text string of summary of page
}
\description{
Get Wikipedia summary
}
