% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{locality_add_habitat_biome}
\alias{locality_add_habitat_biome}
\title{Use azizka/speciesgeocodeR/ and WWF data to encode locations for habitat and biome}
\usage{
locality_add_habitat_biome(locations)
}
\arguments{
\item{locations}{Data.frame containing points (latitude and longitude, perhaps other data as columns)}
}
\value{
data.frame with columns for habitat and biome.
}
\description{
Uses info from http://omap.africanmarineatlas.org/BIOSPHERE/data/note_areas_sp/Ecoregions_Ecosystems/WWF_Ecoregions/WWFecoregions.htm to convert codes to more readable text
}
\examples{
locations <- spocc_taxon_query("Myrmecocystus", limit=50)
locations <- locality_clean(locations)
locations <- locality_add_habitat_biome(locations)
print(head(locations))
}
