% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{locality_clean}
\alias{locality_clean}
\title{Clean locality information}
\usage{
locality_clean(locations)
}
\arguments{
\item{locations}{Data.frame containing points (latitude and longitude, perhaps other data as columns)}
}
\value{
Cleaned data.frame
}
\description{
This uses the ropensci CoordinateCleaner package to clean up points.
}
