% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{match_data}
\alias{match_data}
\title{Get a tree and data, resolve to same set of taxa, and make a shared object}
\usage{
match_data(tree, data, warnings = FALSE)
}
\arguments{
\item{tree}{a phylo object}

\item{data}{A data.frame. Rownames are species names.}

\item{warnings}{Complain about mismatches if TRUE.}
}
\value{
The function returns a list of two elements (phy and data) that are manipulated to include only those species found in both the tree and data supplied by the user. The class of the returned object is "chapter2".
}
\description{
Get a tree and data, resolve to same set of taxa, and make a shared object
}
