% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{pbdb_taxon_query}
\alias{pbdb_taxon_query}
\title{Get information on specimens from the paleobiology data base (largely for extinct taxa)}
\usage{
pbdb_taxon_query(taxon)
}
\arguments{
\item{taxon}{taxon name}
}
\value{
a data frame of name, paleo longitude, paleo latitude, earliest record in ma (minma), latest record in ma (maxma), earliest interval, and latest interval
}
\description{
Get information on specimens from the paleobiology data base (largely for extinct taxa)
}
