% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{run_biogeobears}
\alias{run_biogeobears}
\title{Run biogeobears analyses}
\usage{
run_biogeobears(phy, locations)
}
\arguments{
\item{phy}{The phylo object}

\item{locations}{The states for biogeographic regions}
}
\value{
A list with fits for each model
}
\description{
Run a set of biogeobears analyses to estimate parameters of biogeographic models
}
