% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{run_chapter2}
\alias{run_chapter2}
\title{Do chapter2 analysis}
\usage{
run_chapter2(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
A list with all output
}
\description{
Do chapter2 analysis
}
