% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{render_chapter2}
\alias{render_chapter2}
\title{Create a file of results}
\usage{
render_chapter2(taxon, format = "pdf")
}
\arguments{
\item{taxon}{Clade of interest}

\item{format}{Format: pdf or html}
}
\value{
Nothing, though a file is created in the current working directory
}
\description{
Create a file of results
}
