% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R, R/runanalysis.R
\name{run_biogeobears}
\alias{run_biogeobears}
\title{Continuous to discrete}
\usage{
run_biogeobears(phy, locations)

run_biogeobears(phy, locations)
}
\arguments{
\item{phy}{The phylo object}

\item{locations}{The states for biogeographic regions}

\item{latlon}{The data.frame of points, with species, latitude, and longitude columns}
}
\value{
A data.frame as above but with a region column added

A list with fits for each model
}
\description{
Convert a set of latitudes and longitudes to biogeographic regions

Run a set of biogeobears analyses to estimate parameters of biogeographic models
}
