% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{spocc_taxon_query}
\alias{spocc_taxon_query}
\title{Query many sources at once using spocc to get latitude and longitude}
\usage{
spocc_taxon_query(taxon, limit = 10000, sources = c("gbif", "inat",
  "idigbio"), has_coords = TRUE, by_species = TRUE, verbose = TRUE,
  ...)
}
\arguments{
\item{taxon}{The taxon to get. If a vector, loops over it.}

\item{limit}{The limit of records per site to get (default is maximum of any site)}

\item{sources}{Vector of sources (see ?spocc::occ)}

\item{has_coords}{Boolean for whether to only return records with longitude and latitude data}

\item{by_species}{Boolean: if TRUE, separates the taxon into species first (and searches for the higher level taxon as well)}

\item{verbose}{Boolean: if TRUE, print out progress}

\item{...}{Other arguments to pass to spocc::occ}
}
\value{
data.frame of results
}
\description{
Query many sources at once using spocc to get latitude and longitude
}
\examples{
locations <- spocc_taxon_query("Myrmecocystus", limit=50)
}
